Fix for: Sorting when external sorting is enabled is broken
Provides fix for the following condition:
When external sorting is enabled, attempting to sort via multiple
columns is futile because as soon as you click the column, all of the
sorting turns off.  The problem is related NG-Grid improperly handling
an unexpected DOM event.  Behavior is unmodified if no externalsorting
is false.